Allow users to remove items from cart
Users can remove items in two ways. The first is to click the 'subtract' icon
until the item quantity is reduced to 0. The second is by clicking the 'X'
remove icon.

To hopefully get the icons working properly, I switched from using the CDN to
installing a material_icons gem. Won't know if this fully solves the Heroku
production issue until we push up.