Update minor fixes from waffle stories
- Currency now displays in view using number_to_currency. Also included those
helper methods in the test-helper so the tests use those too.

- Moved rails_12factor gem to only the production environment

- Added route_downcaser gem to make url case insensitive. May cause issue with
how params come in (one instance of Category title param coming in as lower
case and needed to call #capitalize on it. If this is a regular issue may need
to revert and manually implement a fix for url case sensitivity).

- Centered headers on login and create account views

- Added 'the' to flash message for adding an item to duffel for readability